move sparse_softmax_cross_entropy code from modules to functional
